+++ src/inmanta/protocol/decorators.py
@@ -67,7 +67,7 @@ def method(
     api: bool = None,
     agent_server: bool = False,
     validate_sid: bool = None,
-    client_types: List[str] = ["public"],
+    client_types: List[str] = ["api"],
     api_version: int = 1,
     api_prefix: str = "api",
     envelope: bool = False,
@@ -85,7 +85,10 @@ def method(
         :param server_agent: This is a call from the Server to the Agent (reverse http channel through long poll)
         :param agent_server: This is a call from the Agent to the Server
         :param validate_sid: This call requires a valid session, true by default if agent_server and not api
-        :param client_types: The allowed client types for this call
+        :param client_types: The allowed client types for this call. The valid values are:
+                api: The method can be used as by external clients like the Dashboard, cli or 3rd party services
+                compiler: The method is called by the compiler to communicate with the server
+                agent: The method is called by the agent to communicate with the server
         :param arg_options: Options related to arguments passed to the method. The key of this dict is the name of the arg to
             which the options apply. The value is another dict that can contain the following options:
                 header: Map this argument to a header with the following name.
@@ -133,7 +136,7 @@ def typedmethod(
     api: bool = None,
     agent_server: bool = False,
     validate_sid: bool = None,
-    client_types: List[str] = ["public"],
+    client_types: List[str] = ["api"],
     api_version: int = 1,
     api_prefix: str = "api",
     envelope_key: str = const.ENVELOPE_KEY,
@@ -150,7 +153,10 @@ def typedmethod(
         :param server_agent: This is a call from the Server to the Agent (reverse http channel through long poll)
         :param agent_server: This is a call from the Agent to the Server
         :param validate_sid: This call requires a valid session, true by default if agent_server and not api
-        :param client_types: The allowed client types for this call
+        :param client_types: The allowed client types for this call. The valid values are:
+                api: The method can be used as by external clients like the Dashboard, cli or 3rd party services
+                compiler: The method is called by the compiler to communicate with the server
+                agent: The method is called by the agent to communicate with the server
         :param arg_options: Options related to arguments passed to the method. The key of this dict is the name of the arg to
             which the options apply. The value is another dict that can contain the following options:
                 header: Map this argument to a header with the following name.